+++ asdf/_tests/commands/tests/test_tags.py
@@ -4,6 +4,7 @@ import pytest
 
 from asdf import AsdfFile
 from asdf.commands import list_tags
+from asdf.exceptions import AsdfDeprecationWarning
 
 
 @pytest.mark.parametrize("display_classes", [True, False])
@@ -19,7 +20,8 @@ def test_all_tags_present():
     tags = {line.strip() for line in iostream.readlines()}
 
     af = AsdfFile()
-    for tag in af._type_index._type_by_tag:
-        assert tag in tags
+    with pytest.warns(AsdfDeprecationWarning, match="AsdfFile.type_index is deprecated"):
+        for tag in af.type_index._type_by_tag:
+            assert tag in tags
     for tag in af.extension_manager._converters_by_tag:
         assert tag in tags